backend/requirements/base.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
django-filter>=24.0,<25.0
django-cors-headers>=4.4,<5.0
django-environ>=0.11,<0.12
drf-spectacular>=0.27,<0.28
mysqlclient>=2.2,<2.3
